+++ front/app/containers/Admin/settings/customize/AllInput.tsx
@@ -18,7 +18,7 @@ import { StyledSection, StyledSectionTitle } from './Events';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from './messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';