+++ front/app/containers/Admin/settings/customize/Events.tsx
@@ -19,7 +19,7 @@ import {
 // i18n
 import { InjectedIntlProps } from 'react-intl';
 import { FormattedMessage, injectIntl } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from './messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';