+++ front/app/containers/Admin/settings/customize/index.tsx
@@ -24,7 +24,8 @@ import { isCLErrorJSON } from 'utils/errorUtils';
 // i18n
 import { InjectedIntlProps } from 'react-intl';
 import { injectIntl } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from './messages';
+import sharedSettingsMessages from '../messages';
 
 // services
 import { localeStream } from 'services/locale';
@@ -277,6 +278,7 @@ class SettingsCustomizeTab extends PureComponent<
             latestAppConfigStyleSettings={latestAppConfigStyleSettings}
             latestAppConfigCoreSettings={latestAppConfigCoreSettings}
             setParentState={setState}
+            getSetting={getSetting}
           />
 
           <ProjectHeader
@@ -309,10 +311,10 @@ class SettingsCustomizeTab extends PureComponent<
             loading={this.state.loading}
             status={getSubmitState({ errors, saved, state: this.state })}
             messages={{
-              buttonSave: messages.save,
-              buttonSuccess: messages.saveSuccess,
-              messageError: messages.saveErrorMessage,
-              messageSuccess: messages.saveSuccessMessage,
+              buttonSave: sharedSettingsMessages.save,
+              buttonSuccess: sharedSettingsMessages.saveSuccess,
+              messageError: sharedSettingsMessages.saveErrorMessage,
+              messageSuccess: sharedSettingsMessages.saveSuccessMessage,
             }}
           />
         </form>